# api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: myhwproj@yandex.ru
  description: REST api for MyHwProj
  license:
    name: The MIT License (MIT)
    url: https://mit-license.org/
  title: MyHwProj
  version: 1.0.0
servers:
- url: http://localhost:8080/
...
paths:
  /v1/teacher/homeworks:
    get:
      description: Returns homeworks sorted by deadline
      operationId: getHomeworksTeacher
      parameters:
      - description: The number of items to skip before starting to collect the result
          set
        in: query
        name: offset
        required: true
        schema:
          type: integer
      - description: The numbers of items to return
        in: query
        name: limit
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Homework'
                type: array
          description: Successful operation
      summary: Get homeworks
      tags:
      - teacher
    post:
      description: Homework will be shown to students after publicationDatetime
      operationId: addHomeworkTeacher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewHomework'
        description: Homework that needs to be added to the system
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Homework'
          description: Successful operation
        "400":
          content: {}
          description: Invalid homework
      summary: Add new homework
      tags:
      - teacher
      x-codegen-request-body-name: newHomework
  /v1/teacher/homeworks/{homeworkId}:
    get:
      description: Returns a single homework with specified ID
      operationId: getHomeworkByIdTeacher
      parameters:
      - description: ID of homework
        in: path
        name: homeworkId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Homework'
          description: successful operation
        "400":
          content: {}
          description: Invalid ID supplied
        "404":
          content: {}
          description: Homework not found
      summary: Get homework
      tags:
      - teacher
  /v1/teacher/homeworks/{homeworkId}/submissions:
    get:
      description: Returns submissions of homework with specified id sorted by deadline
      operationId: getHomeworkSubmissionsTeacher
      parameters:
      - description: ID of homework
        in: path
        name: homeworkId
        required: true
        schema:
          format: int64
          type: integer
      - description: The number of items to skip before starting to collect the result
          set
        in: query
        name: offset
        required: true
        schema:
          type: integer
      - description: The numbers of items to return
        in: query
        name: limit
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Submission'
                type: array
          description: Successful operation
        "400":
          content: {}
          description: Invalid ID supplied
        "404":
          content: {}
          description: Homework not found
      summary: Get homework submissions
      tags:
      - teacher
  /v1/teacher/submissions/{submissionId}:
    get:
      description: Returns a single submission with specified id
      operationId: getSubmissionTeacher
      parameters:
      - description: ID of submission
        in: path
        name: submissionId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: Successful operation
        "400":
          content: {}
          description: Invalid ID supplied
        "404":
          content: {}
          description: Submission not found
      summary: Get submission
      tags:
      - teacher
...
  /teacher/homeworks:
    get:
      description: Returns page with homeworks sorted by deadline
      operationId: getHomeworksPageTeacher
      parameters:
      - description: Order of page to return
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            text/html:
              example: <html><body>Some HTML</body></html>
          description: Successful operation
        "400":
          content: {}
          description: Invalid page supplied
        "404":
          content: {}
          description: Page not found
      summary: Get homeworks page
      tags:
      - teacher pages
...
  /teacher/homeworks/{homeworkId}:
    get:
      description: Returns page with homework with specified ID and its submissions
      operationId: getHomeworkPageTeacher
      parameters:
      - description: ID of homework
        in: path
        name: homeworkId
        required: true
        schema:
          format: int64
          type: integer
      - description: Order of page to return
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            text/html:
              example: <html><body>Some HTML</body></html>
          description: Successful operation
        "400":
          content: {}
          description: Invalid ID or page supplied
        "404":
          content: {}
          description: Homework or page not found
      summary: Get homework page
      tags:
      - teacher pages
  /teacher/submissions/{submissionId}:
    get:
      description: Returns page with submission with specified ID
      operationId: getSubmissionPageTeacher
      parameters:
      - description: ID of submission
        in: path
        name: submissionId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            text/html:
              example: <html><body>Some HTML</body></html>
          description: Successful operation
        "400":
          content: {}
          description: Invalid ID supplied
        "404":
          content: {}
          description: Submission not found
      summary: Get submission page
      tags:
      - teacher pages
...
components:
  schemas:
    NewHomework:
      example:
        task: Write our own hwproj
        publicationDatetime: 2000-01-23T04:56:07.000+00:00
        name: MyHwProj Implementation
        check: ./gradlew test
        deadline: 2000-01-23T04:56:07.000+00:00
      properties:
        name:
          description: Name of a new homework
          example: MyHwProj Implementation
          type: string
        publicationDatetime:
          description: Date and time of homework publication to students. If time
            zone is not specified, +3 is taken
          format: date-time
          type: string
        check:
          description: Command to check submissions
          example: ./gradlew test
          type: string
        task:
          description: Description of the task
          example: Write our own hwproj
          type: string
        deadline:
          description: Deadline for students. If time zone is not specified, +3 is
            taken
          format: date-time
          type: string
      required:
      - check
      - deadline
      - name
      - publicationDatetime
      - task
      type: object
    Homework:
      example:
        task: Write our own hwproj
        publicationDatetime: 2000-01-23T04:56:07.000+00:00
        name: MyHwProj Implementation
        id: 0
        check: ./gradlew test
        deadline: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          format: int64
          type: integer
        name:
          description: Name of a new homework
          example: MyHwProj Implementation
          type: string
        publicationDatetime:
          description: Date and time of homework publication to students. If time
            zone is not specified, +3 is taken
          format: date-time
          type: string
        check:
          description: Command to check submissions
          example: ./gradlew test
          type: string
        task:
          description: Description of the task
          example: Write our own hwproj
          type: string
        deadline:
          format: date-time
          type: string
      required:
      - check
      - deadline
      - id
      - name
      - publicationDatetime
      - task
      type: object
...
    Submission:
      example:
        homeworkId: 6
        datetime: 2000-01-23T04:56:07.000+00:00
        solution: https://github.com/KaperD/HSE-SD-Course-Roguelike
        comment: OK
        id: 0
        mark: 1
      properties:
        id:
          format: int64
          type: integer
        homeworkId:
          format: int64
          type: integer
        datetime:
          description: Date and time of making this submission. If time zone is not
            specified, +3 is taken
          format: date-time
          type: string
        solution:
          description: Link to the solution
          example: https://github.com/KaperD/HSE-SD-Course-Roguelike
          type: string
        mark:
          description: Mark out of 10. Null, if there were no checks
          format: int32
          type: integer
        comment:
          description: Message from checker
          example: OK
          type: string
      required:
      - comment
      - datetime
      - homeworkId
      - id
      - mark
      - solution
      type: object
...